refactor: rename decompressGzip to gzipMetrics and remove redundant Buffer.from
The middleware only records metrics (no decompression logic), so the file
name now reflects its actual purpose. Also removes an unnecessary
Buffer.from() wrapper since gzipSync accepts strings directly.
